Renamed Data.getData() methods to more descriptive names.
git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/sphinx4@352 94700074-3cef-4d97-a70e-9c8c206c02f5

+++ edu/cmu/sphinx/frontend/SpectrumAnalyzer.java
@@ -86,8 +86,7 @@ public class SpectrumAnalyzer extends DataProcessor {
 
 	SphinxProperties properties = SphinxProperties.getSphinxProperties("");
 
-        float sampleRate = properties.getFloat
-            (FrontEnd.PROP_SAMPLE_RATE, 8000.0F);
+        int sampleRate = properties.getInt(FrontEnd.PROP_SAMPLE_RATE, 8000);
 
         float windowSizeInMs = properties.getFloat
             (FrontEnd.PROP_WINDOW_SIZE_MS, 25.625F);
